% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/english_education.R
\name{english_education_files}
\alias{english_education_files}
\alias{english_education}
\title{TidyTuesday's English Education Data}
\usage{
english_education_files()
}
\value{
\code{character(2)} with file names.
}
\description{
These two files describe the 'Educational attainment of young
people in English towns' from the the \href{https://www.ons.gov.uk/}{UK Office for National Statistics}. It was explored in the July
2023 article
\href{https://www.ons.gov.uk/peoplepopulationandcommunity/educationandchildcare/articles/whydochildrenandyoungpeopleinsmallertownsdobetteracademicallythanthoseinlargertowns/2023-07-25}{"Why do children and young people in smaller towns do better academically than those in larger towns?"}.

Two files are available:
\itemize{
\item \code{english_education.csv} was prepared as part of the \href{https://github.com/rfordatascience/tidytuesday}{Tidy Tuesday} series
on \href{https://github.com/rfordatascience/tidytuesday/blob/master/data/2024/2024-01-23/readme.md}{Educational attainment of young people in English towns}
(2024-01-23). The page also describes the variables.
\item \code{edu_income_eprivation_and_educational_attainment.csv} was
downloaded from the UK Office for National Statistics and
converted from xls to csv. The variables in this table are also
described on the page linked above.
}
}
\examples{

english_education_files()
}
\references{
\itemize{
\item \href{https://www.ons.gov.uk/peoplepopulationandcommunity/educationandchildcare/articles/whydochildrenandyoungpeopleinsmallertownsdobetteracademicallythanthoseinlargertowns/2023-07-25}{"Why do children and young people in smaller towns do better academically than those in larger towns?"}.
\item Tidy Tuesday data from 2024-01-23: https://github.com/rfordatascience/tidytuesday/blob/master/data/2024/2024-01-23/readme.md
}
}
\author{
Laurent Gatto
}
