% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faillites.R
\name{faillites_be}
\alias{faillites_be}
\alias{faillites}
\title{Evolution mensuelle des faillites par NACE}
\usage{
faillites_be()
}
\value{
\code{character(2)} with file names.
}
\description{
Ce jeu de données reprednt l'évolution mensuelle des faillites par
NACE - 15 jours de 2005 à 2023, telles que distribuées par
statbel, l'office belge de statistique.
}
\section{Traitement des données}{

\itemize{
\item Les données ont été téléchargées de
\url{https://statbel.fgov.be/fr/open-data/evolution-mensuelle-des-faillites-par-nace}.
\item Un sous-échantillon de 11544 observation a été sélectionné, de
telle manière a garder les entreprises de grandes tailles. Voir
le \code{scripts/faillites.R}.
}
}

\section{Ce package re-distribute deux fichiers}{

\itemize{
\item \code{TF_BANKRUPTCIES_subset.txt.zip}: échantillon des données sous
format compressé, les valeurs étant séparées pas le caractère
\code{|} (format identique aux données complètes distribuées par
statbel).
\item \code{Method_BANKRUPTCIES.xlsx}: les métadonnées décrivant les
variables, en format xlsx, tel que distribué par statbel.
}
}

\examples{

faillites_be()
}
