% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_exam_data}
\alias{load_exam_data}
\title{Make and load exam data}
\usage{
load_exam_data(noma)
}
\arguments{
\item{noma}{`character(1)` with the student number. Must be
coercible to a `numeric`.}
}
\value{
Invisibly returns `TRUE`. Used for its side effect to load
    an object of class `MSnSet` in the user's global environment.
}
\description{
This function generates data based on a student number and loads
it in the users environment.
}
\examples{
load_exam_data("0123")
x0123
}
