\name{mvylng}
\alias{mvylng}

\docType{data}

\title{
  Data from Mulvey et al. 2015
}

\description{

  This is the "mulvey2015" data from Mulvey et al., _Dynamic proteomic
  profiling of extra-embryonic endoderm differentiation in mouse
  embryonic stem cells. _, Stem Cell. (PMID 26059426). See below for
  more details.

  It was extracted from \code{pRolocdata} package and converted into a
  long format.


}

\usage{data("mvylng")}

\examples{
data(mvylng)
mvylng
}

\keyword{datasets}
