\name{population}
\alias{population_be.csv}

\docType{data}

\title{Belgium population}

\description{

  A set of csv files providing the population sizes in different regions
  in Belgium from 1970, 1981 and 1991 to 2023.

}

\usage{
population_be.csv()
}


\examples{
population_be.csv()

read.csv(population_be.csv()[1])

}

\keyword{datasets}
