% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbim1207.R
\docType{package}
\name{rWSBIM1207}
\alias{rWSBIM1207}
\alias{rWSBIM1207-package}
\title{rWSBIM1207: Companion package for WSBIM1207 course}
\description{
This package is used to distribute data and general documentation about the
WSBIM1207 course, and to make course administration easier. For details about
the course and the course material, see
\url{http://bit.ly/WSBIM1207}.
}
\section{rWSBIM1207 datasets}{


The following data sets are available. Consult the respective manual pages
for further details.
\itemize{
\item data 1: the first data, ...
\item data 2: the second data, ...
}
}

