% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valentine.R
\name{valentine}
\alias{valentine}
\alias{valentin}
\alias{historical_spending}
\alias{gifts_age}
\alias{gifts_gender}
\title{TidyTuesday's Valentine's Day Consumer Data}
\usage{
valentine()
}
\value{
\code{character(3)} of urls to file names.
}
\description{
These data, originally organized by Suraj Das for a Kaggle
dataset, have been cleaned and prepared for the \href{https://github.com/rfordatascience/tidytuesday/blob/master/data/2024/2024-02-13/readme.md}{TidyTuesday series}.

The data contains three files:
\itemize{
\item \code{historical_spending} contains Valentine's spending from 2010 to 2022.
\item \code{gifts_age} contains information on gifts by age.
\item \code{gifts_gender} contains information on gifts by gender.
}

For a description of the variables of these table, see the link
above.

Note that in these historical surveys, gender was collected as
only 'Men' and 'Women', which does not accurately include all
genders.
}
\examples{

valentine()
}
\author{
Laurent Gatto
}
