\name{kem}
\alias{kem}
\alias{kem_count}
\alias{kem_annot}
\alias{kem.tsv}

\docType{data}

\title{
  KEM RNA-Seq data
}


\description{

  This data is composed of two files, namely \code{kem_counts.tsv},
  containing the counts data and the annotation file
  \code{kem_annot.tsv} containing the annotations the KEM samples. Both
  files are encoded as tab-separtated sheets and can be found with the
  \code{kem.tsv()} function.

  Credit: the data have been generated by Mr Kevin Missault.

}


\format{
  
  The RNA-Seq count data for 13 ENSEMBL transcripts (ref) and 16 KEM
  samples is encoded as:
  
  \describe{
    \item{\code{ref}}{ENSEMBL transcript identifiers.}
    \item{\samp{KEM182-01}}{Expression counts for all genes in sample KEM182-01.}
    \item{\samp{KEM182-02}}{Expression counts for all genes in sample KEM182-02.}
    \item{\samp{KEM182-03}}{Expression counts for all genes in sample KEM182-03.}
    \item{\samp{KEM182-04}}{Expression counts for all genes in sample KEM182-04.}
    \item{\samp{KEM182-05}}{Expression counts for all genes in sample KEM182-05.}
    \item{\samp{KEM182-06}}{Expression counts for all genes in sample KEM182-06.}
    \item{\samp{KEM182-07}}{Expression counts for all genes in sample KEM182-07.}
    \item{\samp{KEM182-08}}{Expression counts for all genes in sample KEM182-08.}
    \item{\samp{KEM182-09}}{Expression counts for all genes in sample KEM182-09.}
    \item{\samp{KEM182-10}}{Expression counts for all genes in sample KEM182-10.}
    \item{\samp{KEM182-11}}{Expression counts for all genes in sample KEM182-11.}
    \item{\samp{KEM182-12}}{Expression counts for all genes in sample KEM182-12.}
    \item{\samp{KEM182-13}}{Expression counts for all genes in sample KEM182-13.}
    \item{\samp{KEM182-14}}{Expression counts for all genes in sample KEM182-14.}
    \item{\samp{KEM182-15}}{Expression counts for all genes in sample KEM182-15.}
    \item{\samp{KEM182-16}}{Expression counts for all genes in sample KEM182-16.}
  }

  The annotation contains the following variables for the 16 observations:
  
  \describe{
    \item{\code{sample_id}}{Sample identifier.}
    \item{\code{jurkat}}{A character (yes or no) defining of the cells a Jurkat cells.}
    \item{\code{cell_type}}{Jurkat cell type (A or B).}
    \item{\code{treatment}}{Treatment: either none or stimulated.}
  }  
}


\examples{
kem.tsv()
}

\keyword{datasets}
